requirements.txt
facenet-pytorch == 2.6.0
numpy~=1.26.1
pillow~=10.2.0
matplotlib~=3.8.0
tqdm~=4.67.0